requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2025.4
Django==5.1.7
django-auto-logout==0.5.1
django-crispy-forms==2.4
sqlparse==0.5.3
gunicorn
djangorestframework
drf-spectacular
